+++ src/main/resources/db/migration/V3__UTILITIES.sql
@@ -2,5 +2,5 @@ CREATE TABLE utilities
 (
     id    INT PRIMARY KEY,
     name  VARCHAR(100)   NOT NULL,
-    price NUMERIC(10, 2) NOT NULL
+    price BIGINT NOT NULL
 );
